refactor: extract SettingsPermissionsPane.swift
Extracts the SettingsPermissionsPane View (82 lines) from SettingsAudioPanes.swift
into its own file.

Applied memory rule (range preservation): initial extraction started too early
at line 60 (SettingsRecordingAudioPane's closing brace); corrected to line 63
(SettingsPermissionsPane's doc comment start).

Also added AVFoundation import (needed for AVCaptureDevice references).

Reduces SettingsAudioPanes.swift from 147 to 60 lines.

Tests: 667 (unchanged).

⚠️ Manual smoke check recommended before merge: render Settings > Permissions
pane and verify layout / behavior unchanged.
